refactor: 사다리 생성 역할 분리
