+++ src/test/java/ladder/domain/generator/LadderGeneratorTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.generator;
 
 import static ladder.domain.Direction.RIGHT;
 import static ladder.domain.Direction.STAY;
@@ -6,27 +6,32 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
-import ladder.domain.generator.DirectionGenerator;
-import ladder.domain.generator.LineGenerator;
-import ladder.domain.generator.TestDirectionGenerator;
+import ladder.domain.Direction;
+import ladder.domain.Height;
+import ladder.domain.Ladder;
+import ladder.domain.Players;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class LadderTest {
+class LadderGeneratorTest {
 
     @Test
     @DisplayName("사다리 높이만큼 선을 생성한다.")
     void validLadder() {
+        //given
         final ArrayList<Direction> directions = Lists.newArrayList(
                 RIGHT, STAY, STAY, RIGHT, RIGHT, STAY, RIGHT, STAY, RIGHT, RIGHT);
-        final DirectionGenerator directionGenerator = new TestDirectionGenerator(directions);
-        final LineGenerator lineGenerator = new LineGenerator(directionGenerator);
+        final var directionGenerator = new TestDirectionGenerator(directions);
+        final var lineGenerator = new LineGenerator(directionGenerator);
+        final var ladderGenerator = new LadderGenerator(lineGenerator);
         final Players players = new Players(List.of("pobi", "crong"));
         final Height height = new Height(5);
 
-        final Ladder ladder = new Ladder(lineGenerator, players, height);
+        //when
+        Ladder ladder = ladderGenerator.generate(players, height);
 
+        //then
         assertThat(ladder.getLines()).hasSize(5);
     }
 }
